requirements/dev.txt
amqp==5.2.0
asgiref==3.8.1
astroid==2.15.8
asttokens==2.4.1
attrs==23.2.0
bandit==1.7.8
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==5.1.1
Django==4.2.11
django-celery-results==2.5.1
django-ckeditor-5==0.2.12
django-cleanup==8.1.0
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-guardian==2.4.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-taggit==5.0.1
executing==2.0.1
flake8==6.1.0
flake8-annotations==3.1.1
flake8-bandit==4.1.1
flake8-bugbear==24.4.26
flake8-comprehensions==3.14.0
flake8-django==1.4
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
flake8-quotes==3.4.0
flake8-variables-names==0.0.6
ipython==8.23.0
jedi==0.19.1
kombu==5.3.7
lazy-object-proxy==1.10.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
packaging==24.0
parso==0.8.4
pbr==6.0.0
pexpect==4.9.0
pillow==10.3.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.1.0
Pygments==2.17.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
rich==13.7.1
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
stevedore==5.2.0
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
Unidecode==1.3.8
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
